+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraAudioChannel.kt
@@ -20,13 +20,16 @@ class SoraAudioChannel(
         private val channelId:         String,
         private val signalingMetadata: String = "",
         private var streamType:        SoraStreamType,
-        private var codec:             SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS,
+        private var audioCodec:        SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS,
+        private val audioBitRate:      Int? = null,
         private var sdpSemantics:      PeerConnection.SdpSemantics =
                 PeerConnection.SdpSemantics.UNIFIED_PLAN,
         private var listener:          Listener?
 ) {
 
-    val TAG = SoraAudioChannel::class.simpleName
+    companion object {
+        private val TAG = SoraAudioChannel::class.simpleName
+    }
 
     interface Listener {
         fun onConnect(channel: SoraAudioChannel) {}
@@ -88,7 +91,9 @@ class SoraAudioChannel(
                 enableMultistream()
             }
 
-            audioCodec = this@SoraAudioChannel.codec
+            audioCodec = this@SoraAudioChannel.audioCodec
+            audioBitrate = this@SoraAudioChannel.audioBitRate
+
             sdpSemantics = this@SoraAudioChannel.sdpSemantics
         }