+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraVideoChannel.kt
@@ -16,6 +16,8 @@ import jp.shiguredo.sora.sdk.error.SoraErrorReason
 import jp.shiguredo.sora.sdk.util.SoraLogger
 import org.webrtc.*
 import android.os.Handler
+import jp.shiguredo.sora.sample.stats.VideoUpstreamLatencyStatsCollector
+import jp.shiguredo.sora.sdk.channel.option.PeerConnectionOption
 
 class SoraVideoChannel(
         private val context:                 Context,
@@ -30,11 +32,13 @@ class SoraVideoChannel(
         private var videoEnabled:            Boolean = true,
         private val videoWidth:              Int = SoraVideoOption.FrameSize.Portrait.VGA.x,
         private val videoHeight:             Int = SoraVideoOption.FrameSize.Portrait.VGA.y,
+        private val simulcast:               Boolean = false,
         private val videoFPS:                Int =  30,
         private val fixedResolution:         Boolean = false,
         private val videoCodec:              SoraVideoOption.Codec = SoraVideoOption.Codec.VP9,
         private val audioCodec:              SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS,
-        private val videoBitrate:            Int? = null,
+        private val videoBitRate:            Int? = null,
+        private val audioBitRate:            Int? = null,
         private val needLocalRenderer:       Boolean = true,
         private val audioEnabled:            Boolean = true,
         private val sdpSemantics:            PeerConnection.SdpSemantics =
@@ -45,7 +49,7 @@ class SoraVideoChannel(
 ) {
 
     companion object {
-        val TAG = SoraVideoChannel::class.simpleName
+        private val TAG = SoraVideoChannel::class.simpleName
     }
 
     private var egl: EglBase? = EglBase.create()
@@ -59,6 +63,8 @@ class SoraVideoChannel(
         fun onAddLocalRenderer(channel: SoraVideoChannel, renderer: SurfaceViewRenderer) {}
         fun onAttendeesCountUpdated(channel: SoraVideoChannel, attendees: ChannelAttendeesCount) {}
     }
+
+    private val statsCollector = VideoUpstreamLatencyStatsCollector()
 
     private val channelListener = object : SoraMediaChannel.Listener {
 
@@ -75,7 +81,14 @@ class SoraVideoChannel(
         }
 
         override fun onError(mediaChannel: SoraMediaChannel, reason: SoraErrorReason) {
-            SoraLogger.d(TAG, "[video_channel] @onError")
+            SoraLogger.d(TAG, "[video_channel] @onError $reason")
+            handler.post {
+                listener?.onError(this@SoraVideoChannel, reason)
+            }
+        }
+
+        override fun onError(mediaChannel: SoraMediaChannel, reason: SoraErrorReason, message: String) {
+            SoraLogger.d(TAG, "[video_channel] @onError $reason: $message")
             handler.post {
                 listener?.onError(this@SoraVideoChannel, reason)
             }
@@ -132,6 +145,12 @@ class SoraVideoChannel(
             SoraLogger.d(TAG, "[video_channel] @onPushMessage ${push}")
         }
 
+        override fun onPeerConnectionStatsReady(mediaChannel: SoraMediaChannel, statsReport: RTCStatsReport) {
+            // statsReport.statsMap.entries.forEach {
+            //     SoraLogger.d(TAG, "${it.key}=${it.value}")
+            // }
+            statsCollector.newStatsReport(statsReport)
+        }
     }
 
     private var mediaChannel:  SoraMediaChannel? = null
@@ -197,11 +216,33 @@ class SoraVideoChannel(
                 enableMultistream()
             }
 
+            if(this@SoraVideoChannel.simulcast) {
+                enableSimulcast()
+                // hardware encoder では動かせていない、ソフトウェアを指定する
+                videoEncoderFactory = SoftwareVideoEncoderFactory()
+            }
             spotlight    = this@SoraVideoChannel.spotlight
             videoCodec   = this@SoraVideoChannel.videoCodec
+            videoBitrate = this@SoraVideoChannel.videoBitRate
+
             audioCodec   = this@SoraVideoChannel.audioCodec
-            videoBitrate = this@SoraVideoChannel.videoBitrate
+            audioBitrate = this@SoraVideoChannel.audioBitRate
+
+            // 全部デフォルト値なので、実際には指定する必要はない
+            audioOption = SoraAudioOption().apply {
+                useHardwareAcousticEchoCanceler = true
+                useHardwareNoiseSuppressor      = true
+
+                audioProcessingEchoCancellation = true
+                audioProcessingAutoGainControl  = true
+                audioProcessingHighpassFilter   = true
+                audioProcessingNoiseSuppression = true
+            }
             sdpSemantics = this@SoraVideoChannel.sdpSemantics
+        }
+
+        val peerConnectionOption = PeerConnectionOption().apply {
+            getStatsIntervalMSec = 5000
         }
 
         mediaChannel = SoraMediaChannel(
@@ -212,7 +253,8 @@ class SoraVideoChannel(
                 signalingNotifyMetadata = signalingNotifyMetatada,
                 mediaOption             = mediaOption,
                 listener                = channelListener,
-                clientId                = clientId
+                clientId                = clientId,
+                peerConnectionOption    = peerConnectionOption
         )
         mediaChannel!!.connect()
     }