+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/VideoChatRoomActivity.kt
@@ -1,6 +1,7 @@
 package jp.shiguredo.sora.sample.ui
 
 import android.annotation.TargetApi
+import android.content.Context
 import android.content.res.Resources
 import android.graphics.Color
 import android.media.AudioManager
@@ -30,7 +31,7 @@ import java.util.*
 class VideoChatRoomActivity : AppCompatActivity() {
 
     companion object {
-        val TAG = VideoChatRoomActivity::class.simpleName
+        private val TAG = VideoChatRoomActivity::class.simpleName
     }
 
     private var channelName = ""
@@ -39,13 +40,17 @@ class VideoChatRoomActivity : AppCompatActivity() {
     private var videoCodec:  SoraVideoOption.Codec = SoraVideoOption.Codec.VP9
     private var audioCodec:  SoraAudioOption.Codec = SoraAudioOption.Codec.OPUS
     private var audioEnabled = true
-    private var bitRate: Int? = null
+    private var audioBitRate: Int? = null
+    private var videoBitRate: Int? = null
     private var videoWidth: Int = SoraVideoOption.FrameSize.Portrait.VGA.x
     private var videoHeight: Int = SoraVideoOption.FrameSize.Portrait.VGA.y
+    private var simulcast = false
     private var fps: Int = 30
     private var fixedResolution = false
     private var clientId: String? = null
     private var sdpSemantics = PeerConnection.SdpSemantics.UNIFIED_PLAN
+
+    private var oldAudioMode: Int = AudioManager.MODE_INVALID
 
     private var streamType = SoraStreamType.BIDIRECTIONAL
 
@@ -56,7 +61,7 @@ class VideoChatRoomActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setupWindow()
 
-        channelName = intent.getStringExtra("CHANNEL_NAME")
+        channelName = intent.getStringExtra("CHANNEL_NAME") ?: getString(R.string.channelId)
 
         spotlight = intent.getIntExtra("SPOTLIGHT", 0)
 
@@ -66,9 +71,9 @@ class VideoChatRoomActivity : AppCompatActivity() {
             else  -> true
         }
 
-        videoCodec = SoraVideoOption.Codec.valueOf(intent.getStringExtra("VIDEO_CODEC"))
+        videoCodec = SoraVideoOption.Codec.valueOf(intent.getStringExtra("VIDEO_CODEC") ?: "VP9")
 
-        audioCodec = SoraAudioOption.Codec.valueOf(intent.getStringExtra("AUDIO_CODEC"))
+        audioCodec = SoraAudioOption.Codec.valueOf(intent.getStringExtra("AUDIO_CODEC") ?: "OPUS")
 
         streamType = when (intent.getStringExtra("STREAM_TYPE")) {
             "BIDIRECTIONAL" -> SoraStreamType.BIDIRECTIONAL
@@ -84,7 +89,7 @@ class VideoChatRoomActivity : AppCompatActivity() {
             else  -> true
         }
 
-        fps = intent.getStringExtra("FPS").toInt()
+        fps = (intent.getStringExtra("FPS") ?: "30").toInt()
 
         var videoSize = when (intent.getStringExtra("VIDEO_SIZE")) {
             // Portrait
@@ -107,15 +112,25 @@ class VideoChatRoomActivity : AppCompatActivity() {
         videoWidth = videoSize.x
         videoHeight = videoSize.y
 
+        simulcast = when (intent.getStringExtra("SIMULCAST")) {
+            "ENABLED" -> true
+            else      -> false
+        }
+
         fixedResolution = when (intent.getStringExtra("RESOLUTION_CHANGE")) {
             "VARIABLE" -> false
             "FIXED"    -> true
             else       -> false
         }
 
-        bitRate = when (intent.getStringExtra("BITRATE")) {
+        videoBitRate = when (intent.getStringExtra("VIDEO_BIT_RATE")) {
             "UNDEFINED" -> null
-            else -> intent.getStringExtra("BITRATE").toInt()
+            else -> intent.getStringExtra("VIDEO_BIT_RATE")?.toInt()
+        }
+
+        audioBitRate = when (intent.getStringExtra("AUDIO_BIT_RATE")) {
+            "UNDEFINED" -> null
+            else -> intent.getStringExtra("AUDIO_BIT_RATE")?.toInt()
         }
 
         clientId = when (intent.getStringExtra("CLIENT_ID")) {
@@ -163,11 +178,20 @@ class VideoChatRoomActivity : AppCompatActivity() {
     override fun onResume() {
         super.onResume()
         this.volumeControlStream = AudioManager.STREAM_VOICE_CALL
+        val audioManager = applicationContext.getSystemService(Context.AUDIO_SERVICE)
+                as AudioManager
+        oldAudioMode = audioManager.mode
+        Log.d(TAG, "AudioManager mode change: ${oldAudioMode} => MODE_IN_COMMUNICATION(3)")
+        audioManager.mode = AudioManager.MODE_IN_COMMUNICATION
     }
 
     override fun onPause() {
         Log.d(TAG, "onPause")
         super.onPause()
+        val audioManager = applicationContext.getSystemService(Context.AUDIO_SERVICE)
+                as AudioManager
+        Log.d(TAG, "AudioManager mode change: MODE_IN_COMMUNICATION(3) => ${oldAudioMode}")
+        audioManager.mode = oldAudioMode
         close()
     }
 
@@ -224,12 +248,14 @@ class VideoChatRoomActivity : AppCompatActivity() {
                 videoEnabled      = videoEnabled,
                 videoWidth        = videoWidth,
                 videoHeight       = videoHeight,
+                simulcast         = simulcast,
                 videoFPS          = fps,
                 fixedResolution   = fixedResolution,
                 videoCodec        = videoCodec,
-                videoBitrate      = bitRate,
+                videoBitRate      = videoBitRate,
                 audioEnabled      = audioEnabled,
                 audioCodec        = audioCodec,
+                audioBitRate      = audioBitRate,
                 sdpSemantics      = sdpSemantics,
                 streamType        = streamType,
                 clientId          = clientId,